minimal/* entry point
