+++ packages/waku/package.json
@@ -42,6 +42,14 @@
       "types": "./dist/client.d.ts",
       "default": "./dist/client.js"
     },
+    "./minimal/server": {
+      "types": "./dist/minimal/server.d.ts",
+      "default": "./dist/minimal/server.js"
+    },
+    "./minimal/client": {
+      "types": "./dist/minimal/client.d.ts",
+      "default": "./dist/minimal/client.js"
+    },
     "./server": {
       "types": "./dist/server.d.ts",
       "default": "./dist/server.js"
